chore: move report to demand-order
